+++ worker/src/mails_api/index.ts
@@ -9,6 +9,7 @@ import auto_reply from './auto_reply'
 import webhook_settings from './webhook_settings';
 import s3_attachment from './s3_attachment';
 import address_auth from './address_auth';
+import { ensureDefaultSendBalance } from './send_mail_api';
 
 export const api = new Hono<HonoCustomType>()
 
@@ -96,7 +97,13 @@ api.get('/api/settings', async (c) => {
 
     const no_limit_roles = getSplitStringListValue(c.env.NO_LIMIT_SEND_ROLE);
     const is_no_limit_send_balance = user_role && no_limit_roles.includes(user_role);
-    const balance = is_no_limit_send_balance ? 99999 : await c.env.DB.prepare(
+    const noLimitSendAddressList = is_no_limit_send_balance ?
+        [] : await getJsonSetting(c, CONSTANTS.NO_LIMIT_SEND_ADDRESS_LIST_KEY) || [];
+    const isNoLimitSendAddress = noLimitSendAddressList?.includes(address);
+    if (!is_no_limit_send_balance && !isNoLimitSendAddress) {
+        await ensureDefaultSendBalance(c, address);
+    }
+    const balance = (is_no_limit_send_balance || isNoLimitSendAddress) ? 99999 : await c.env.DB.prepare(
         `SELECT balance FROM address_sender where address = ? and enabled = 1`
     ).bind(address).first("balance");
     return c.json({